requirements.txt
torchxrayvision
torchvision
pandas
requests
umap-learn
notebook
seaborn
openpyxl
torchmetrics
pillow
numpy
scikit-image
scikit-learn
tqdm
PySide6
pydicom
